Total gasto na loja done

+++ BeFeira/Shared/Cliente.cs
@@ -9,7 +9,7 @@ namespace BeFeira.Shared
         public string Username { get; set; } = string.Empty;
         public string Password { get; set; } = string.Empty;
         public string Email { get; set; } = string.Empty;
-        public int Total { get; set; } = 0;
+        public float Total { get; set; } = 0;
 
         public string? UrlProfilePic { get; set; } = "https://cdn-icons-png.flaticon.com/512/5087/5087579.png";
